requirements.txt
certifi==2018.4.16
chardet==3.0.4
Django==2.1
django-extensions==2.1.0
django-rest==0.0.1
djangorestframework==3.8.2
idna==2.7
jsonfield==2.0.2
pytz==2018.5
requests==2.19.1
six==1.11.0
urllib3==1.23
